requirements.txt
aiogram==3.13.1
aiohttp==3.9.5
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.0.2
pytest-asyncio==0.23.5
